Crate atomic_waker

futures::task::AtomicWaker extracted into its own crate.

Features

This crate adds a feature, portable-atomic, which uses a polyfill from the portable-atomic crate in order to provide functionality to targets without atomics. See the README for the portable-atomic crate for more information on how to use it.

Structs

AtomicWaker

A synchronization primitive for task wakeup.